sepehrsmartservices/Alaedin/requirements.txt
django_environ==0.12.0
Flask==3.0.0
jdatetime==5.2.0
lxml==5.3.2
redis==5.2.1
Requests==2.32.3
gunicorn==23.0.0
selenium==4.31.0
blinker==1.7.0
selenium-wire==5.1.0
